requirements.txt
Flask
psycopg2-binary
pandas
flask>=2.3.0
werkzeug>=2.3.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
python-dotenv>=1.0.0
flask>=2.3.0
werkzeug>=2.3.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0flask-cors>=4.0.0
